Split out deleteTheme
